Refactor morph selector to reusable component
Extracted the morph selection UI and logic from create-product-view and edit-product-view into a new MorphSelector component for reuse and maintainability. Updated imports and type definitions accordingly. Also renamed DeleteProductModal.tsx to delete-product-modal.tsx for consistency, and refactored Supabase middleware to simplify session handling.
