+++ src/features/product/components/create-product-view.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useActionState, useEffect, useState, useMemo } from "react";
+import { useActionState, useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import { createProductAction } from "@/features/product/product.actions";
 import type { CreateProductState } from "@/features/product/product.types";
@@ -9,6 +9,7 @@ import { useAppToast } from "@/hooks/useAppToast";
 import { ImageUploader } from "@/components/ui/ImageUploader";
 import { Button } from "@/components/ui/Button";
 import { Breadcrumb } from "@/components/ui/Breadcrumb";
+import { MorphSelector, type SelectedMorph } from "./morph-selector";
 
 // --- Type Definitions ---
 type Morph = Tables<"morphs">;
@@ -17,43 +18,6 @@ type MorphCategory = Tables<"morph_categories"> & {
   morphs: Morph[];
   sub_categories: MorphSubCategory[];
 };
-type SelectedMorph = Morph & { color_hex?: string };
-
-// --- SVG Icons ---
-const PlusIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <line x1="12" y1="5" x2="12" y2="19"></line>
-    <line x1="5" y1="12" x2="19" y2="12"></line>
-  </svg>
-);
-const CloseIcon = ({ ...props }) => (
-  <svg
-    {...props}
-    xmlns="http://www.w3.org/2000/svg"
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    {" "}
-    <line x1="18" y1="6" x2="6" y2="18"></line>{" "}
-    <line x1="6" y1="6" x2="18" y2="18"></line>{" "}
-  </svg>
-);
 
 interface CreateProductViewProps {
   allMorphs: MorphCategory[];
@@ -72,8 +36,6 @@ export function CreateProductView({ allMorphs }: CreateProductViewProps) {
 
   // State for managing UI that isn't part of the form submission
   const [selectedMorphs, setSelectedMorphs] = useState<SelectedMorph[]>([]);
-  const [searchText, setSearchText] = useState("");
-  const [isInputFocused, setIsInputFocused] = useState(false);
   const [pictures, setPictures] = useState<File[]>([]);
   const [picturePreviews, setPicturePreviews] = useState<string[]>([]);
 
@@ -87,15 +49,9 @@ export function CreateProductView({ allMorphs }: CreateProductViewProps) {
     if (!selectedMorphs.find((m) => m.id === morph.id)) {
       let color_hex: string | undefined;
       for (const cat of allMorphs) {
-        if (cat.morphs?.some((m) => m.id === morph.id)) {
-          color_hex = cat.color_hex ?? undefined;
-          break;
-        }
+        if (cat.morphs?.some((m) => m.id === morph.id)) { color_hex = cat.color_hex ?? undefined; break; }
         for (const sub of cat.sub_categories ?? []) {
-          if (sub.morphs?.some((m) => m.id === morph.id)) {
-            color_hex = sub.color_hex ?? undefined;
-            break;
-          }
+          if (sub.morphs?.some((m) => m.id === morph.id)) { color_hex = sub.color_hex ?? undefined; break; }
         }
         if (color_hex) break;
       }
@@ -103,40 +59,9 @@ export function CreateProductView({ allMorphs }: CreateProductViewProps) {
     }
   };
 
-  const handleRemoveMorph = (morphToRemove: Morph) => {
+  const handleRemoveMorph = (morphToRemove: SelectedMorph) => {
     setSelectedMorphs(selectedMorphs.filter((m) => m.id !== morphToRemove.id));
   };
-
-  const filteredCategories = useMemo(() => {
-    const text = searchText.toLowerCase();
-    const selectedIds = new Set(selectedMorphs.map((m) => m.id));
-    const result: MorphCategory[] = [];
-    for (const cat of allMorphs) {
-      const matchedMorphs = (cat.morphs ?? []).filter(
-        (m) =>
-          (!text || m.name.toLowerCase().includes(text)) &&
-          !selectedIds.has(m.id)
-      );
-      const matchedSubs = (cat.sub_categories ?? [])
-        .map((sub) => ({
-          ...sub,
-          morphs: (sub.morphs ?? []).filter(
-            (m) =>
-              (!text || m.name.toLowerCase().includes(text)) &&
-              !selectedIds.has(m.id)
-          ),
-        }))
-        .filter((sub) => sub.morphs.length > 0);
-      if (matchedMorphs.length > 0 || matchedSubs.length > 0) {
-        result.push({
-          ...cat,
-          morphs: matchedMorphs,
-          sub_categories: matchedSubs,
-        });
-      }
-    }
-    return result;
-  }, [searchText, allMorphs, selectedMorphs]);
 
   const generateYearOptions = () => {
     const currentYear = new Date().getFullYear();
@@ -239,131 +164,13 @@ export function CreateProductView({ allMorphs }: CreateProductViewProps) {
           )}
         </div>
 
-        <div>
-          <label className="block text-sm font-semibold text-gray-700 mb-1">
-            Genetics (Morphs)*
-          </label>
-          {selectedMorphs.length > 0 && (
-            <div
-              className={`flex flex-wrap gap-2 rounded-lg border p-2 min-h-[4rem] mb-2 ${
-                state.errors.morphs ? "border-red-500" : "border-gray-300"
-              }`}
-            >
-              {selectedMorphs.map((morph) => (
-                <div
-                  key={morph.id}
-                  className="flex items-center gap-2 bg-white border border-gray-300 rounded-full px-3 py-1 text-sm font-medium text-gray-700"
-                >
-                  <input type="hidden" name="morphs" value={morph.id} />
-                  <span
-                    className="h-2 w-2 rounded-full"
-                    style={{ backgroundColor: morph.color_hex ?? "#9CA3AF" }}
-                  ></span>
-                  <span>{morph.name}</span>
-                  <button
-                    type="button"
-                    onClick={() => handleRemoveMorph(morph)}
-                  >
-                    <CloseIcon className="h-3 w-3 text-gray-500 hover:text-gray-800" />
-                  </button>
-                </div>
-              ))}
-            </div>
-          )}
-          {state.errors.morphs && (
-            <p className="mt-1 text-sm text-red-600">
-              {state.errors.morphs[0]}
-            </p>
-          )}
-
-          <div className="relative w-full mt-2">
-            <input
-              type="text"
-              placeholder="Add"
-              value={searchText}
-              onChange={(e) => setSearchText(e.target.value)}
-              onFocus={() => setIsInputFocused(true)}
-              onBlur={() => setTimeout(() => setIsInputFocused(false), 150)}
-              className="py-2.5 ps-4 pe-10 block w-full border-gray-300 rounded-lg text-sm focus:border-blue-500 focus:ring-blue-500"
-            />
-            <div className="absolute top-1/2 right-3 -translate-y-1/2 pointer-events-none text-gray-400">
-              <PlusIcon />
-            </div>
-            {isInputFocused && (
-              <div className="absolute z-50 w-full max-h-72 mt-1 p-1 bg-white border border-gray-200 rounded-lg overflow-y-auto shadow">
-                {filteredCategories.length === 0 && (
-                  <div className="px-4 py-2 text-sm text-gray-500">
-                    No results found.
-                  </div>
-                )}
-                {filteredCategories.map((category) => (
-                  <div key={category.id}>
-                    <div className="flex items-center gap-x-2 px-3 py-2">
-                      <span
-                        className="font-semibold text-sm"
-                        style={{ color: "#9CA3AF" }}
-                      >
-                        {category.name}
-                      </span>
-                      <span className="flex-grow border-t border-dashed border-gray-300 dark:border-neutral-700"></span>
-                    </div>
-                    {(category.morphs ?? []).map((morph) => (
-                      <button
-                        type="button"
-                        key={morph.id}
-                        onClick={() => {
-                          handleAddMorph(morph);
-                          setSearchText("");
-                        }}
-                        className="flex items-center gap-x-2 w-full text-left text-sm px-4 py-2 hover:bg-gray-100 rounded-md"
-                      >
-                        <span
-                          className="h-2 w-2 rounded-full"
-                          style={{
-                            backgroundColor: category.color_hex ?? "#9CA3AF",
-                          }}
-                        ></span>
-                        <span>{morph.name}</span>
-                      </button>
-                    ))}
-                    {(category.sub_categories ?? []).map((sub) => (
-                      <div key={sub.id} className="pl-3">
-                        <div className="flex items-center gap-x-2 px-3 py-1">
-                          <span
-                            className="text-xs font-medium"
-                            style={{ color: "#9CA3AF" }}
-                          >
-                            {sub.name}
-                          </span>
-                          <span className="flex-grow border-t border-dashed border-gray-300 dark:border-neutral-700"></span>
-                        </div>
-                        {(sub.morphs ?? []).map((morph) => (
-                          <button
-                            type="button"
-                            key={morph.id}
-                            onClick={() => {
-                              handleAddMorph(morph);
-                              setSearchText("");
-                            }}
-                            className="flex items-center gap-x-2 w-full text-left text-sm px-4 py-2 hover:bg-gray-100 rounded-md"
-                          >
-                            <span
-                              className="h-2 w-2 rounded-full"
-                              style={{
-                                backgroundColor: sub.color_hex ?? "#9CA3AF",
-                              }}
-                            ></span>
-                            <span>{morph.name}</span>
-                          </button>
-                        ))}
-                      </div>
-                    ))}
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        </div>
+        <MorphSelector
+          allMorphs={allMorphs}
+          selectedMorphs={selectedMorphs}
+          onAddMorph={handleAddMorph}
+          onRemoveMorph={handleRemoveMorph}
+          error={state.errors.morphs?.[0]}
+        />
 
         <div className="grid grid-cols-2 gap-4">
           <div>
