+++ src/middleware.ts
@@ -1,15 +1,10 @@
 import { type NextRequest } from 'next/server'
-import { createClient } from '@/lib/supabase/middleware'
+import { updateSession } from '@/lib/supabase/middleware'
 
 export async function middleware(request: NextRequest) {
-  // สร้าง client และ response จาก helper ที่เราสร้างไว้
-  const { supabase, response } = createClient(request)
-
-  // คำสั่งนี้สำคัญมาก: ทำหน้าที่ refresh session ของ user
-  // เพื่อให้ Server Components และ Server Actions รู้จัก user ที่ล็อกอินอยู่เสมอ
-  await supabase.auth.getSession()
-
-  return response
+  // ฟังก์ชัน updateSession นี้จะจัดการเรื่องการรีเฟรช session cookie
+  // และทำให้ Server Components รู้จัก user ที่ล็อกอินอยู่เสมอ
+  return await updateSession(request)
 }
 
 export const config = {
@@ -19,6 +14,7 @@ export const config = {
      * - _next/static (static files)
      * - _next/image (image optimization files)
      * - favicon.ico (favicon file)
+     * Feel free to modify this pattern to include more paths.
      */
     '/((?!_next/static|_next/image|favicon.ico).*)',
   ],
